all id's changed to long

+++ src/main/java/com/f90782/cars/data/entity/Document.java
@@ -20,7 +20,7 @@ import lombok.Setter;
 public class Document {
   @Id
   @GeneratedValue
-  private int idDocument;
+  private long idDocument;
   @OneToOne
   @JoinColumn(name = "id_reservation")
   private Reservation idReservation;
