+++ src/main/java/com/f90782/cars/data/entity/Shop.java
@@ -17,6 +17,6 @@ import lombok.Setter;
 public class Shop {
   @Id
   @GeneratedValue
-  private int idShop;
+  private long idShop;
   private String shopName;
 }
